+++ src/roq/kraken/order_entry.cpp
@@ -28,11 +28,6 @@ static const auto SUPPORTS = utils::Mask{
     SupportType::CANCEL_ORDER,
     SupportType::ORDER_ACK,
 };
-static const auto SUPPORTS_MASTER = utils::Mask{
-    SUPPORTS,
-    SupportType::REFERENCE_DATA,
-    SupportType::MARKET_STATUS,
-};
 
 static const auto ALLOW_PIPELINING = true;
 
@@ -47,10 +42,9 @@ OrderEntry::OrderEntry(
     core::io::Context &context,
     uint16_t stream_id,
     Security &security,
-    Shared &shared,
-    bool master)
+    Shared &shared)
     : handler_(handler), stream_id_(stream_id),
-      name_(fmt::format("{}:{}:{}"sv, stream_id_, NAME, security.get_account())), master_(master),
+      name_(fmt::format("{}:{}:{}"sv, stream_id_, NAME, security.get_account())),
       connection_(
           *this,
           context,
@@ -68,14 +62,10 @@ OrderEntry::OrderEntry(
           .disconnect = create_metrics(name_, "disconnect"sv),
       },
       profile_{
-          .assets = create_metrics(name_, "assets"sv),
-          .assets_ack = create_metrics(name_, "assets_ack"sv),
-          .asset_pairs = create_metrics(name_, "asset_pairs"sv),
-          .asset_pairs_ack = create_metrics(name_, "asset_pairs_ack"sv),
-          .positions = create_metrics(name_, "positions"sv),
-          .positions_ack = create_metrics(name_, "positions_ack"sv),
           .get_web_sockets_token = create_metrics(name_, "get_web_sockets_token"sv),
           .get_web_sockets_token_ack = create_metrics(name_, "get_web_sockets_token_ack"sv),
+          .positions = create_metrics(name_, "positions"sv),
+          .positions_ack = create_metrics(name_, "positions_ack"sv),
       },
       latency_{
           .ping = create_metrics(name_, "ping"sv),
@@ -101,14 +91,10 @@ void OrderEntry::operator()(metrics::Writer &writer) {
       // counter
       .write(counter_.disconnect, metrics::COUNTER)
       // profile
-      .write(profile_.assets, metrics::PROFILE)
-      .write(profile_.assets_ack, metrics::PROFILE)
-      .write(profile_.asset_pairs, metrics::PROFILE)
-      .write(profile_.asset_pairs_ack, metrics::PROFILE)
-      .write(profile_.positions, metrics::PROFILE)
-      .write(profile_.positions_ack, metrics::PROFILE)
       .write(profile_.get_web_sockets_token, metrics::PROFILE)
       .write(profile_.get_web_sockets_token_ack, metrics::PROFILE)
+      .write(profile_.positions, metrics::PROFILE)
+      .write(profile_.positions_ack, metrics::PROFILE)
       // latency
       .write(latency_.ping, metrics::LATENCY);
 }
@@ -148,7 +134,7 @@ void OrderEntry::operator()(ConnectionStatus status) {
     StreamStatus stream_status{
         .stream_id = stream_id_,
         .account = security_.get_account(),
-        .supports = (master_ ? SUPPORTS_MASTER : SUPPORTS).get(),
+        .supports = SUPPORTS.get(),
         .status = status_,
         .type = StreamType::REST,
         .priority = Priority::PRIMARY,
@@ -192,20 +178,6 @@ uint32_t OrderEntry::download(OrderEntryState state) {
     case OrderEntryState::TOKEN:
       get_token();
       return 1;
-    case OrderEntryState::ASSETS:
-      if (master_) {
-        get_assets();
-        return 1;
-      } else {
-        return {};
-      }
-    case OrderEntryState::ASSET_PAIRS:
-      if (master_) {
-        get_asset_pairs();
-        return 1;
-      } else {
-        return {};
-      }
     case OrderEntryState::POSITIONS:
       get_positions();
       return 1;
@@ -287,182 +259,6 @@ void OrderEntry::operator()(const server::Trace<json::Token> &event) {
       .token = token.token,
   };
   handler_(token_update);
-}
-
-// assets
-
-void OrderEntry::get_assets() {
-  profile_.assets([&]() {
-    auto method = core::http::Method::GET;
-    auto path = "/0/public/Assets"sv;
-    core::web::Request request{
-        .method = method,
-        .path = path,
-        .query = {},
-        .accept = core::http::Accept::JSON,
-        .content_type = {},
-        .headers = {},
-        .body = {},
-        .quality_of_service = {},
-    };
-    auto sequence = download_.sequence();
-    connection_(
-        "assets"sv, request, [this, sequence]([[maybe_unused]] auto &request_id, auto &response) {
-          auto trace_info = server::create_trace_info();
-          server::Trace event(trace_info, response);
-          get_assets_ack(event, sequence);
-        });
-  });
-}
-
-void OrderEntry::get_assets_ack(
-    const server::Trace<core::web::Response> &event, uint32_t sequence) {
-  profile_.assets_ack([&]() {
-    auto &[trace_info, response] = event;
-    auto state = OrderEntryState::ASSETS;
-    try {
-      auto [status, category, body] = response.result();
-      log::debug(R"(status={}, category={}, body="{}")"sv, status, category, body);
-      if (download_.skip(sequence, state)) {
-        log::info("Download state={} has already been processed"sv, state);
-        return;
-      }
-      response.expect(core::http::Status::OK);
-      core::json::Buffer buffer(decode_buffer_);
-      auto assets = core::json::Parser::create<json::Assets>(body, buffer);
-      server::Trace event(trace_info, assets);
-      (*this)(event);
-      download_.check(state);
-    } catch (core::NetworkError &e) {
-      log::warn(R"(Exception type={}, what="{}")"sv, typeid(e).name(), e.what());
-      download_.retry(state);
-    }
-  });
-}
-
-void OrderEntry::operator()(const server::Trace<json::Assets> &event) {
-  auto &[trace_info, assets] = event;
-  log::info<4>("assets={}"sv, assets);
-  // do nothing
-}
-
-// asset-pairs
-
-void OrderEntry::get_asset_pairs() {
-  profile_.asset_pairs([&]() {
-    auto method = core::http::Method::GET;
-    auto path = "/0/public/AssetPairs"sv;
-    core::web::Request request{
-        .method = method,
-        .path = path,
-        .query = {},
-        .accept = core::http::Accept::JSON,
-        .content_type = {},
-        .headers = {},
-        .body = {},
-        .quality_of_service = {},
-    };
-    auto sequence = download_.sequence();
-    connection_(
-        "asset_pairs"sv,
-        request,
-        [this, sequence]([[maybe_unused]] auto &request_id, auto &response) {
-          auto trace_info = server::create_trace_info();
-          server::Trace event(trace_info, response);
-          get_asset_pairs_ack(event, sequence);
-        });
-  });
-}
-
-void OrderEntry::get_asset_pairs_ack(
-    const server::Trace<core::web::Response> &event, uint32_t sequence) {
-  profile_.asset_pairs_ack([&]() {
-    auto &[trace_info, response] = event;
-    auto state = OrderEntryState::ASSET_PAIRS;
-    try {
-      auto [status, category, body] = response.result();
-      log::debug(R"(status={}, category={}, body="{}")"sv, status, category, body);
-      if (download_.skip(sequence, state)) {
-        log::info("Download state={} has already been processed"sv, state);
-        return;
-      }
-      response.expect(core::http::Status::OK);
-      core::json::Buffer buffer(decode_buffer_);
-      auto asset_pairs = core::json::Parser::create<json::AssetPairs>(body, buffer);
-      server::Trace event(trace_info, asset_pairs);
-      (*this)(event);
-      download_.check(state);
-    } catch (core::NetworkError &e) {
-      log::warn(R"(Exception type={}, what="{}")"sv, typeid(e).name(), e.what());
-      download_.retry(state);
-    }
-  });
-}
-
-void OrderEntry::operator()(const server::Trace<json::AssetPairs> &event) {
-  auto &[trace_info, asset_pairs] = event;
-  log::info<4>("asset_pairs={}"sv, asset_pairs);
-  assert(std::empty(asset_pairs.error));
-  std::vector<std::string> symbols;
-  symbols.reserve(std::size(asset_pairs.result));
-  size_t counter = {};
-  for (auto &item : asset_pairs.result) {
-    log::info<2>("item={}"sv, item);
-    if (std::empty(item.wsname)) {
-      log::info<1>(R"(Skipping altname="{}", reason: wsname is empty)"sv, item.altname);
-      continue;
-    }
-    std::string symbol(item.wsname);
-    // remove escape
-    symbol.erase(std::remove(std::begin(symbol), std::end(symbol), '\\'), std::end(symbol));
-    if (shared_.discard_symbol(symbol))
-      continue;
-    if (all_symbols_.emplace(symbol).second)  // only include new
-      symbols.emplace_back(symbol);
-    ++counter;
-    auto tick_size = std::pow(double{10.0}, -static_cast<double>(item.pair_decimals));
-    auto min_trade_vol = std::pow(double{10.0}, -static_cast<double>(item.lot_decimals));
-    ReferenceData reference_data{
-        .stream_id = stream_id_,
-        .exchange = Flags::exchange(),
-        .symbol = symbol,
-        .description = item.altname,
-        .security_type = {},
-        .base_currency = item.aclass_base,
-        .quote_currency = item.aclass_quote,
-        .margin_currency = {},
-        .commission_currency = item.aclass_base,
-        .tick_size = tick_size,
-        .multiplier = item.lot_multiplier,  // XXX check
-        .min_trade_vol = min_trade_vol,
-        .max_trade_vol = NaN,
-        .trade_vol_step_size = min_trade_vol,
-        .option_type = {},
-        .strike_currency = {},
-        .strike_price = NaN,
-        .underlying = {},
-        .time_zone = {},
-        .issue_date = {},
-        .settlement_date = {},
-        .expiry_datetime = {},
-        .expiry_datetime_utc = {},
-    };
-    server::create_trace_and_dispatch(handler_, trace_info, reference_data, true);
-    MarketStatus market_status{
-        .stream_id = stream_id_,
-        .exchange = Flags::exchange(),
-        .symbol = symbol,
-        .trading_status = TradingStatus::OPEN,  // XXX doesn't exist?
-    };
-    server::create_trace_and_dispatch(handler_, trace_info, market_status, true);
-  }
-  log::info("AssetPairs {} / {}"sv, counter, std::size(asset_pairs.result));
-  if (!std::empty(symbols)) {
-    SymbolsUpdate symbols_update{
-        .symbols = symbols,
-    };
-    handler_(symbols_update);
-  }
 }
 
 // positions